Refactor/cleanup scheduler SPI
 - Move scheduler packages into `org.springframework.cloud.deployer.spi.scheduler`
 - Remove RelaxedNames

Resolves #319
