+++ src/test/java/org/springframework/cloud/deployer/spi/scheduler/kubernetes/KubernetesSchedulerPropertiesTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.deployer.scheduler.spi.kubernetes;
+package org.springframework.cloud.deployer.spi.scheduler.kubernetes;
 
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
